requirements.txt
aiohttp
python-dotenv
line-bot-sdk==3.1.0
openai